start big refactor
