+++ templates/template1.ts
@@ -1,4 +1,4 @@
-import { circuit } from "lib/builder/legacy-circuit"
+import { circuit } from "lib/builder"
 
 /**
  * ```
