+++ tests/mergeCircuits1.test.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "bun:test"
-import { CircuitBuilder } from "lib/builder/legacy-circuit"
+import { CircuitBuilder } from "lib/builder"
 import { mergeCircuits } from "lib/expanding/mergeCircuits"
 
 test("mergeCircuits1", () => {
